LJI-3 task. Change the grouping classes strategy by layer strategy. Relocate LoginController to controller package. Delete login package.
